# authorization_service/auth-service/docs/openapi_auth_api_doc.yaml
openapi: 3.0.0
info:
  title: Auth API
  version: 1.0.0
paths:
...
  /login:
      post:
        tags:
          - Authentication
        summary: User authentication endpoint
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
        responses:
          '200':
            description: Successfully authenticated
            headers:
              Set-Cookie:
                schema:
                  type: string
                  example:
                    access_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZXN1bHQiOiJZb3UgYXJlIGF3ZXNvbWUhIn0.PhRXjIVL1yUhAND4uiE-p6V2rXHQ0drCj9156thJAJg; HttpOnly;refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZXN1bHQiOiJZb3UgYXJlIHZlcnkgc21hcnQhIn0.GZvDoQdT9ldwmlPOrZWrpiaHas0DiFmZlytr1dhaxi4; HttpOnly
          '422':
            description: Validation error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/validation-error-resp'
          '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/bad-request-login-resp'
  /logout:
      post:
        tags:
          - Authentication
        summary: User logout endpoint
        security:
          - Cookie: [access_token=]

        responses:
          '204':
            description:
          '401':
            description: Unauthorised
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/unauthorised-resp'
  /token-refresh:
        post:
          tags:
            - Authentication
          summary: Refresh access_token
          security:
            - Cookie: [refresh_token=]

          responses:
            '200':
              description: Ok
              headers:
                Set-Cookie:
                  schema:
                    type: string
                    example:
                      access_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZXN1bHQiOiJZb3UgYXJlIGF3ZXNvbWUhIn0.PhRXjIVL1yUhAND4uiE-p6V2rXHQ0drCj9156thJAJg; HttpOnly;refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZXN1bHQiOiJZb3UgYXJlIHZlcnkgc21hcnQhIn0.GZvDoQdT9ldwmlPOrZWrpiaHas0DiFmZlytr1dhaxi4; HttpOnly
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
            '400':
              description: Bad request
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/bad-request-resp'
            '500':
              description: Internal server error
  /account/{user_id}:
      get:
        tags:
          - Personal account
        summary: User account endpoint
        parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        security:
          - Cookie: [access_token=]

        responses:
          '200':
            description: Ok
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/user-account-info-resp'
          '401':
            description: Unauthorised
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/unauthorised-resp'
  /account/{user_id}/history:
        get:
          tags:
            - Personal account
          summary: History of user logins
          parameters:
          - name: user_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/user-login-history-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
...
components:
  schemas:
...
    validation-error-resp:
      type: object
      properties:
        result:
          type: string
          example: 'validation error'
...
    bad-request-login-resp:
      type: object
      properties:
        result:
          type: string
          example: 'bad request'
        error:
          type: string
          example: 'wrong credentials | inactive user | user not verified'
    bad-request-resp:
      type: object
      properties:
        result:
          type: string
          example: 'bad request'
...
    unauthorised-resp:
      type: object
      properties:
        result:
          type: string
          example: 'unauthorised'
    user-account-info-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        data:
          type: string
          example: "{}"
        id:
          type: string
          example: 'b92ef010-5e4c-4fd0-99d6-41b6456272cd'
        email:
          type: string
          example: 'user@email.com'
    user-login-history-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        data:
          type: array
          example: [{"timestamp": "", "ip_address": "","location": "", "user_agent": ""}]
...